APPLICATION(STREAMLIT)/requirements.txt
streamlit>=1.31.0
numpy>=1.23
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
groq>=0.9.0
